requirements_pro.txt
# PROFESSIONAL CLIMATE-ALPHA REQUIREMENTS
# Install with: pip install -r requirements_pro.txt --break-system-packages

# Core Scientific Computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0

# Deep Learning (optional but recommended)
# tensorflow>=2.8.0  # Uncomment if you want to use TensorFlow models

# Time Series & Statistics
statsmodels>=0.13.0
hmmlearn>=0.2.7

# Technical Analysis
# Option 1: If you have TA-Lib C library installed
# python-talib>=0.4.24

# Option 2: If you don't have TA-Lib C library (Windows users)
talib-binary>=0.4.24

# Data Collection
yfinance>=0.2.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Optional: Jupyter for notebooks
jupyter>=1.0.0
ipython>=7.0.0

# Optional: Performance optimization
numba>=0.55.0

# Optional: Parallel processing
joblib>=1.0.0
